+++ tools/cmf-webpack-plugin/src/index.js
@@ -1,6 +1,7 @@
 /* eslint-disable global-require */
 const path = require('path');
 const mergeSettings = require('@talend/scripts-cmf/cmf-settings.merge');
+const RawSource = require('webpack-sources').RawSource;
 
 /**
  * React CMF Webpack Plugin
@@ -49,7 +50,7 @@ ReactCMFWebpackPlugin.prototype.apply = function reactCMFWebpackPluginApply(comp
 	 * Runs at each webpack run.
 	 * Calls cmf merge function with the modified cmf config.
 	 */
-	function emit(compilation, callback) {
+	function generateCMFSettings(compilation, callback) {
 		this.log(
 			'emit',
 			JSON.stringify({ canRun: this.canRun, lastRun: this.lastRun, lastWatch: this.lastWatch }),
@@ -60,42 +61,22 @@ ReactCMFWebpackPlugin.prototype.apply = function reactCMFWebpackPluginApply(comp
 		this.lastRun = startTime;
 		this.canRun = false;
 
-		this.modifiedFiles = mergeSettings(this.options, callback);
+		this.mergedSettingsObjects = mergeSettings(this.options, () => {}, false);
+
+		this.mergedSettingsObjects.destination.forEach(obj => {
+			compilation.emitAsset(path.basename(obj.path), new RawSource(JSON.stringify(obj.content)));
+		});
+		this.mergedSettingsObjects.sources.forEach(filename => {
+			compilation.fileDependencies.add(filename);
+		});
 
 		const endTime = Date.now();
 		this.log(`Files merged in ${((endTime - startTime) % 60000) / 1000}s`);
 		this.canRun = true;
-
-		callback();
-	}
-
-	/**
-	 * Register the cmf settings files in the watched files.
-	 * So every change will trigger a new webpack run.
-	 */
-	function afterEmit(compilation, callback) {
-		let compilationFileDependencies;
-		if (Array.isArray(compilation.fileDependencies)) {
-			compilationFileDependencies = new Set(compilation.fileDependencies);
-		} else {
-			compilationFileDependencies = compilation.fileDependencies;
-		}
-
-		this.modifiedFiles
-			.filter(file => !compilationFileDependencies.has(file))
-			.forEach(file => compilation.fileDependencies.add(file));
-		if (!compilationFileDependencies.has(cmfconfigPath)) {
-			compilation.fileDependencies.add(cmfconfigPath);
-		}
-		callback();
 	}
 
 	const plugin = { name: 'ReactCMFPlugin' };
-	compiler.hooks.emit.tapAsync(plugin, emit.bind(this));
-
-	if (this.options.watch) {
-		compiler.hooks.afterEmit.tapAsync(plugin, afterEmit.bind(this));
-	}
+	compiler.hooks.thisCompilation.tap(plugin, generateCMFSettings.bind(this));
 };
 
 module.exports = ReactCMFWebpackPlugin;