+++ tools/scripts-config-react-webpack/config/webpack.config.umd.js
@@ -7,7 +7,7 @@ const CircularDependencyPlugin = require('circular-dependency-plugin');
 
 const cdn = require('@talend/scripts-config-cdn');
 
-const exists = require('@talend/scripts-utils/fs');
+const utils = require('@talend/scripts-utils');
 
 const {
 	getCommonStyleLoaders,
@@ -23,10 +23,11 @@ module.exports = options => {
 	const dcwpConfig = options.getUserConfig('dynamic-cdn-webpack-plugin');
 	const cssModulesEnabled = options.getUserConfig(['css', 'modules'], true);
 	const userCopyConfig = options.getUserConfig('copy', []);
-	const useTypescript = exists.tsConfig();
+	const useTypescript = utils.fs.tsConfig();
 	const userSassData = options.getUserConfig('sass', {});
 	const sassData = getSassData(userSassData);
 	const isEnvProd = options.mode === 'production';
+
 	return (env = {}) => {
 		const name = (env && env.umd) || 'Talend';
 		return {
@@ -58,11 +59,11 @@ module.exports = options => {
 					},
 					{
 						test: /\.scss$/,
-						use: getSassLoaders(cssModulesEnabled, sassData, options.mode),
+						use: getSassLoaders(cssModulesEnabled, sassData, false),
 					},
 					{
 						test: /\.css$/,
-						use: getCommonStyleLoaders(false, options.mode),
+						use: getCommonStyleLoaders(false, false),
 					},
 					...getAssetsRules(false),
 				],