+++ tools/upgrade-deps/src/npm.js
@@ -1,15 +1,15 @@
 /* eslint-disable no-console */
 /* eslint-disable no-param-reassign */
 /* eslint-disable no-await-in-loop, no-restricted-syntax */
-const fs = require('fs');
-const fsprom = require('fs/promises');
-const os = require('os');
-const util = require('util');
-const path = require('path');
-const { exec } = require('child_process');
-const semver = require('semver');
-const stripAnsi = require('strip-ansi');
-const colors = require('./colors');
+import fs from 'fs';
+import fsprom from 'fs/promises';
+import os from 'os';
+import util from 'util';
+import path from 'path';
+import { exec } from 'child_process';
+import semver from 'semver';
+import stripAnsi from 'strip-ansi';
+import colors from './colors.js';
 
 const execProm = util.promisify(exec);
 const CWD = process.cwd();
@@ -272,8 +272,7 @@ async function removeFromLockFile(opts) {
 		console.error(e);
 	}
 }
-
-module.exports = {
+export default {
 	checkPackageJson,
 	getUpdate,
 	createPackageJsonManager,