+++ tools/scripts-config-react-webpack/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @talend/scripts-config-react-webpack
 
+## 15.2.6
+
+### Patch Changes
+
+- 464f7207d: Webpack react config - Fix sass HMR on file change for dev mode
+
+## 15.2.5
+
+### Patch Changes
+
+- 093ce8530: fix: add meta in head if inject is false
+- Updated dependencies [c7152471f]
+  - @talend/scripts-utils@2.0.2
+
+## 15.2.4
+
+### Patch Changes
+
+- 65c8ad05c: fix: apply breaking changes on talend/scripts-utils
+- Updated dependencies [65c8ad05c]
+  - @talend/scripts-utils@2.0.0
+
+## 15.2.3
+
+### Patch Changes
+
+- Updated dependencies [c18aabb97]
+  - @talend/scripts-config-babel@12.0.0
+
 ## 15.2.2
 
 ### Patch Changes